Moving to much smaller version (nested TokenSets)
